refactor: PR 반영

+++ src/main/java/lotto/model/LottoRank.java
@@ -33,7 +33,7 @@ public enum LottoRank {
         return this == SECOND;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
